wired fullscreen button and added empty state for offline feed
