+++ package.json
@@ -104,6 +104,7 @@
     "redux-thunk": "^2.3.0",
     "resolve": "^1.20.0",
     "save": "^2.4.0",
+    "screenfull": "^5.1.0",
     "tailwindcss": "^1.9.6",
     "tslib": "^2.2.0",
     "typescript": "^4.3.2",
