+++ src/Components/Facility/ConsultationDetails.tsx
@@ -90,21 +90,6 @@ export const ConsultationDetails = (props: any) => {
       selected === true ? "border-primary-500 text-primary-600 border-b-2" : ""
     }`;
 
-  const viewOptions = [
-    "Patient Head",
-    "Monitor 0",
-    "Upper Body",
-    "Monitor 1",
-    "Lower Body",
-    "Monitor 2",
-    "Monitor 4",
-    "Monitor 5",
-    "Preset 1",
-    "Preset 2",
-    "Preset 3",
-    "Preset 4",
-  ];
-
   return (
     <div>
       {isPrintMode ? (
@@ -417,66 +402,7 @@ export const ConsultationDetails = (props: any) => {
               </div>
             </div>
           )}
-          {tab === "FEED" && (
-            <div className="mt-4 px-6 mb-20">
-              <PageTitle title="Live Feed" hideBack={true} />
-              <div className="mt-4 flex flex-col">
-                <div className="mt-4 flex flex-col md:flex-row">
-                  <LiveFeed></LiveFeed>
-                  <div className="grid grid-cols-2 md:ml-12 my-auto gap-4 mt-4 md:mt-0">
-                    {viewOptions.map((option: any) => (
-                      <div>
-                        <button className="bg-green-100 border border-white rounded-md px-3 py-2 text-black font-semibold hover:bg-green-200 w-full">
-                          {option}
-                        </button>
-                      </div>
-                    ))}
-                  </div>
-                </div>
-
-                <div className="flex flex-row justify-between">
-                  <div className="mt-5 p-2 flex flex-row bg-green-100 border border-white rounded w-9/12 md:w-5/12 justify-evenly">
-                    <div>
-                      <button className="bg-green-100 hover:bg-green-200 border border-green-100 rounded p-2">
-                        <i className="fas fa-arrow-up md:p-2"></i>
-                      </button>
-                    </div>
-                    <div>
-                      <button className="bg-green-100 hover:bg-green-200 border border-green-100 rounded p-2">
-                        <i className="fas fa-arrow-down md:p-2"></i>
-                      </button>
-                    </div>
-                    <div>
-                      <button className="bg-green-100 hover:bg-green-200 border border-green-100 rounded p-2">
-                        <i className="fas fa-arrow-left md:p-2"></i>
-                      </button>
-                    </div>
-                    <div>
-                      <button className="bg-green-100 hover:bg-green-200 border border-green-100 rounded p-2">
-                        <i className="fas fa-arrow-right md:p-2"></i>
-                      </button>
-                    </div>
-                    <div>
-                      <button className="bg-green-100 hover:bg-green-200 border border-green-100 rounded p-2">
-                        <i className="fas fa-search-plus md:p-2"></i>
-                      </button>
-                    </div>
-                    <div>
-                      <button className="bg-green-100 hover:bg-green-200 border border-green-100 rounded p-2">
-                        <i className="fas fa-redo md:p-2"></i>
-                      </button>
-                    </div>
-                  </div>
-
-                  <div className="flex flex-col justify-center mt-5 mr-4 md:mt-auto md:mr-0">
-                    <button>
-                      <i className="fas fa-expand hover:text-black"></i>
-                    </button>
-                  </div>
-                </div>
-              </div>
-            </div>
-          )}
+          {tab === "FEED" && <LiveFeed />}
           {tab === "SUMMARY" && (
             <div className="mt-4">
               <PageTitle title="Primary Parameters Plot" hideBack={true} />
